# lib_ui/src/opeapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Book Library Management API",
    "description": "Backend API documentation for E-Commerce Application",
    "contact": {
      "name": "Meet Yadav",
      "url": "http://github.com/meet",
      "email": "meetyadav78692@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Server"
    },
    {
      "url": "https://api.myproductiondomain.com",
      "description": "Production Server"
    }
  ],
...
  "paths": {
...
    "/notification": {
      "post": {
        "tags": [
          "notification-controller"
        ],
        "operationId": "postAnnouncement",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnouncementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementResponse"
                }
              }
            }
          }
        }
      }
    },
    "/common/signIn": {
      "post": {
        "tags": [
          "Common Operation"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/common/refreshToken": {
      "post": {
        "tags": [
          "Common Operation"
        ],
        "operationId": "refreshToken",
        "parameters": [
          {
            "name": "refreshToken",
            "in": "cookie",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/common/otp/verify": {
      "post": {
        "tags": [
          "Common Operation"
        ],
        "operationId": "verifyOtp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/common/otp/send-otp": {
      "post": {
        "tags": [
          "Common Operation"
        ],
        "summary": "send-Otp",
        "description": "Give email and then send otp to the email",
        "operationId": "sendOtp",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/common/otp/change_password": {
      "post": {
        "tags": [
          "Common Operation"
        ],
        "summary": "Change Password",
        "description": "Give PasswordChange Data to change the password",
        "operationId": "verifyOtpAndChangePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetData"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/common/login": {
      "post": {
        "tags": [
          "Common Operation"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/common/adminRegister": {
      "post": {
        "tags": [
          "Common Operation"
        ],
        "operationId": "adminRegister",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string",
            "maxLength": 10,
            "minLength": 10
          },
          "password": {
            "type": "string",
            "maxLength": 2147483647,
            "minLength": 8
          },
          "profileImg": {
            "type": "string",
            "format": "binary"
          },
          "userRole": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "STUDENT",
                "ADMIN",
                "TEACHER",
                "LIBRARIAN",
                "GUEST"
              ]
            },
            "maxItems": 2147483647,
            "minItems": 1,
            "uniqueItems": true
          }
        },
        "required": [
          "email",
          "name"
        ]
      },
...
      "AnnouncementRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "NEW_BOOK",
              "BOOK_DUE",
              "MEETING",
              "MILESTONE"
            ]
          }
        },
        "required": [
          "message",
          "title",
          "type"
        ]
      },
      "AnnouncementResponse": {
        "type": "object",
        "properties": {
          "announcementId": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "NEW_BOOK",
              "BOOK_DUE",
              "MEETING",
              "MILESTONE"
            ]
          },
          "read": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "jwt": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        }
      },
      "PasswordResetData": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "otp": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "newPassword",
          "otp"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "maxLength": 12,
            "minLength": 2
          }
        },
        "required": [
          "password"
        ]
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "description": "JWT Authentication",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}